feat(session): make ProfileSession.profile mutable; add updateProfile
The bump-on-write path needs to update the session's in-memory profile
value to match what was just written to profile-index.sqlite, so future
session.profile reads observe the bumped dataFormatVersion through
@Observable. updateProfile preconditions on profile.id equality — the
session is keyed on id in SessionManager.sessions, so an identity swap
would orphan the entry. The nonisolated `id` accessor now reads a stable
`profileID` captured at init time, so it does not have to read the
main-actor-isolated mutable `profile` property.

Also refactors cleanupSync(coordinator:) to take an optional coordinator
so task cancellation runs unconditionally. nil-coordinator builds
(Preview / some tests) previously leaked syncReloadTask, setUpTask,
catalogRefreshTask etc. past session teardown. SessionManager call sites
updated to invoke cleanupSync even when syncCoordinator is nil.

The CloudKit-sync reload debouncing, cleanupSync, and updateProfile are
extracted to ProfileSession+SyncCleanup.swift to keep the main file
under SwiftLint's file_length threshold; the relevant private fields are
loosened to module-internal (matching the pre-existing precedent for
pragmaOptimizeTask et al.) so the extension can manage their lifecycle.

Refs #764.
