+++ MoolahTests/App/ProfileSessionUpdateProfileTests.swift
@@ -1,0 +1,23 @@
+import Foundation
+import Testing
+
+@testable import Moolah
+
+@Suite("ProfileSession — updateProfile")
+@MainActor
+struct ProfileSessionUpdateProfileTests {
+  @Test("updateProfile updates the in-memory profile value")
+  func updatesValue() throws {
+    let containerManager = try ProfileContainerManager.forTesting()
+    let original = Profile(label: "Test")
+    let session = try ProfileSession(
+      profile: original, containerManager: containerManager)
+
+    var bumped = original
+    bumped.dataFormatVersion = 1
+    session.updateProfile(bumped)
+
+    #expect(session.profile.dataFormatVersion == 1)
+    #expect(session.profile.id == original.id)
+  }
+}
